Made incr_decr and append_prepend for the store_t interface.

+++ src/btree/incr_decr.cc
@@ -4,7 +4,7 @@
 
 struct btree_incr_decr_oper_t : public btree_modify_oper_t {
 
-    explicit btree_incr_decr_oper_t(bool increment, unsigned long long delta)
+    explicit btree_incr_decr_oper_t(bool increment, uint64_t delta)
         : increment(increment), delta(delta)
     { }
 
@@ -17,7 +17,7 @@ struct btree_incr_decr_oper_t : public btree_modify_oper_t {
         
         // If we can't parse the key as a number, we fail
         bool valid;
-        unsigned long long number;
+        uint64_t number;
         if (old_value->size < 50) {
             char buffer[50];
             memcpy(buffer, old_value->value(), old_value->size);
@@ -52,7 +52,7 @@ struct btree_incr_decr_oper_t : public btree_modify_oper_t {
         // stack.
 
         valuecpy(&temp_value, old_value);
-        int chars_written = sprintf(temp_value.value(), "%llu", (unsigned long long)number);
+        int chars_written = sprintf(temp_value.value(), "%llu", (long long unsigned)number);
         rassert(chars_written <= MAX_IN_NODE_VALUE_SIZE); // Not really necessary.
         temp_value.value_size(chars_written);
 
@@ -61,7 +61,7 @@ struct btree_incr_decr_oper_t : public btree_modify_oper_t {
     }
 
     bool increment;   // If false, then decrement
-    unsigned long long delta;   // Amount to increment or decrement by
+    uint64_t delta;   // Amount to increment or decrement by
 
     /* Used as temporary storage, so that the value we return from operate() doesn't become invalid
     before run_btree_modify_oper is done with it. */
@@ -73,7 +73,7 @@ struct btree_incr_decr_oper_t : public btree_modify_oper_t {
     store_t::incr_decr_result_t result;
 };
 
-store_t::incr_decr_result_t btree_incr_decr(const btree_key *key, btree_slice_t *slice, bool increment, unsigned long long delta, castime_t castime) {
+store_t::incr_decr_result_t btree_incr_decr(const btree_key *key, btree_slice_t *slice, bool increment, uint64_t delta, castime_t castime) {
     btree_incr_decr_oper_t oper(increment, delta);
     run_btree_modify_oper(&oper, slice, key, castime);
     return oper.result;
